Functions exported from ModifiedHankelFunctionsOfOrderOneThird:

ModifiedHankelFunctionsOfOrderOneThird.modifiedhankel — Method

modifiedhankel(z)

Return $h₁$, $h₂$, $h₁'$, and $h₂'$, the first and second modified Hankel functions of order 1/3 and their derivatives.

These functions solve Stokes' equation $d²u/dz² + zu = 0$ as a power series in $z$ for abs2(z) < 36 and an approximate asymptotic expansion otherwise. The asymptotic solution is necessary because the $z³ⁱ$ in the power series blows up as $i → ∞$.

Warning

For very large arguments z, the exponential function called within the asymptotic expansion may overflow and return Inf, Nan, or 0 values. No warnings will be thrown from this module or Base for finite precision issues.

For more information, see [SCL1945].

See also: powerseries, asymptotic

Examples

julia> h1, h2, h1prime, h2prime = modifiedhankel(complex(2.687, -0.648));

References

Private functions:

ModifiedHankelFunctionsOfOrderOneThird.asymptotic — Method

asymptotic(z)

Return $h₁$, $h₂$, $h₁'$, and $h₂'$, the first and second modified Hankel functions of order 1/3 and their derivatives using an asymptotic expansion, useful for abs2(z) > 36.

For more information, see [SCL1945].

See also: modifiedhankel, powerseries

References

ModifiedHankelFunctionsOfOrderOneThird.powerseries — Method

powerseries(z)

Return $h₁$, $h₂$, $h₁'$, and $h₂'$, the first and second modified Hankel functions of order 1/3 and their derivatives using a power series with 30 terms.

For more information, see [SCL1945].

See also: modifiedhankel, asymptotic

References

SCL1945The Staff of the Computation Library (1945), Tables of the modified Hankel function of order one-third and of their derivatives. Cambridge, MA: Harvard University Press.

SCL1945The Staff of the Computation Library (1945), Tables of the modified Hankel function of order one-third and of their derivatives. Cambridge, MA: Harvard University Press.

SCL1945The Staff of the Computation Library (1945), Tables of the modified Hankel function of order one-third and of their derivatives. Cambridge, MA: Harvard University Press.